spotless and fix 0.8 compatibility
