+++ instrumentation/openai-client-instrumentation/instrumentation-0.8/src/main/java/co/elastic/otel/openai/v0_2/OpenAiOkHttpClientBuilderInstrumentation.java
