+++ instrumentation/openai-client-instrumentation/instrumentation-0.8/src/test/java/co/elastic/otel/openai/v0_2/ChatTest.java
