+++ renovate.json
@@ -19,6 +19,6 @@
     ".buildkite/**",
     "examples/**",
     "jvmti-access/jni-build/*.Dockerfile",
-    "instrumentation/openai-client-instrumentation/instrumentation-0.2/build.gradle.kts"
+    "instrumentation/openai-client-instrumentation/instrumentation-0.8/build.gradle.kts"
   ]
 }
