+++ settings.gradle.kts
@@ -19,7 +19,7 @@ include("custom")
 include("instrumentation")
 include("instrumentation:openai-client-instrumentation:common")
 include("instrumentation:openai-client-instrumentation:testing-common")
-include("instrumentation:openai-client-instrumentation:instrumentation-0.2")
+include("instrumentation:openai-client-instrumentation:instrumentation-0.8")
 include("instrumentation:openai-client-instrumentation:instrumentation-0.14")
 include("inferred-spans")
 include("resources")
